96 make meter api able to save data to the db (#97)
* GC Suppress

* Swap to double and implement snapshot service

* add post put and patch methods

* spacing

* improve gfunction

+++ meter-api/meter-api/Datatypes/Database/MeterAgent.cs
@@ -6,7 +6,7 @@ namespace meter_api.Datatypes.Database
     {
         public required string Id { get; set; }
         public required string DisplayName { get; set; }
-        public required double TotalUsage { get; set; }
+        public required decimal TotalUsage { get; set; }
         public required decimal TotalBilling { get; set; }
     }
 }
